project/requirements.txt
apturl==0.5.2
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
data==0.4
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
deprecation==2.0.7
distro==1.9.0
distutils-pytest==0.2.1
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
evdev==1.7.1
eyed3==0.9.7
feedparser==6.0.10
filelock==3.13.1
filetype==1.2.0
funcsigs==1.0.2
ghp-import==2.1.0
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httplib2==0.20.4
idna==3.10
ifaddr==0.2.0
IMDbPY==2021.4.18
iniconfig==2.0.0
Jinja2==3.1.4
keyboard==0.13.5
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
MouseInfo==0.1.3
msgpack==1.0.3
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.2
onboard==1.4.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
PAM==0.4.2
pandas==2.2.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
psutil==5.9.8
ptyprocess==0.7.0
pyasyncore==1.0.2
PyAutoGUI==0.9.54
pycairo==1.25.1
pycparser==2.22
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.3
pyelftools==0.30
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
pyinotify==0.9.6
PyJWT==2.7.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.7
pynvim==0.5.0
pyparsing==3.1.1
pyparted==3.12.0
PyPDF2==3.0.1
pyperclip==1.9.0
pypng==0.20231004.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.3
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-gnupg==0.5.2
python-magic==0.4.27
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.2
pyyaml_env_tag==0.1
qrcode==7.4.2
redis==5.2.1
repolib==2.2.1
requests==2.32.3
requests-file==1.5.1
rich==13.7.1
selenium==4.25.0
setproctitle==1.3.3
setuptools==75.2.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.3
systemd-python==235
tinycss2==1.2.1
tk==0.1.0
tldextract==3.1.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ufw==0.36.2
undetected-chromedriver==3.5.5
Unidecode==1.3.8
urllib3==2.2.3
uv==0.5.9
wadllib==1.3.6
watchdog==5.0.3
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wheel==0.42.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
yt-dlp==2024.4.9
